Add uniqueness checks to the core graph helpers API. This now throws an IntegrityError if one of the specified unique keys already exists. Adding the IntegrityError to the low-level API has entailed added throws declarations throughout much of the rest of the system.
Adding uniqueness constrains the AccessibleEntity.identifier has
revealed some problems with the REST API tests, and particularly
the need to setup and teardown the database every time. The problem
here is is that this is difficult, and apparently not possible to
restore a fresh DB without restarting the server.

Tests still failing as is.