Initial stage of extreme persistance refactoring.
Validation is now done on the full tree before we attempt to persist
anything (rather than recursively, before save). This makes is less
likely stuff'll end up in the graph that shouldn't be when imports
have some validation errors but ignore them. It also cleans up some
of the tortured logic in the DAO class.

There is still a lot of work to do to rationalise the code, particularly
in the BundleFieldValidator, which is unpleasantly stateful.

All tests pass, but more tests are needed, particularly on createOrUpdate
behaviour.