Basic functionality for item promotion, which allows bypassing access restrictions without removing them entirely. This functionality is primarily intended for use with Links and Annotations (which are currently the only Promotable items), but the API is implemented in a generic manner and could in theory be used elsewhere, perhaps as a way to draw attention to new or interesting items.
RE: the implementation. An additional permission type ('promote') has
been added to distinguish promotion from ordinary setting of visibility
(which requires 'update' - or 'owner') permissions, since the owner of
an item should not necessarily be able to promote items they've created.

More tests need to be added to explore the edge cases and ensure no massive
holes have been put in the ACL.