Make VirtualUnit a DescribedEntity so it can have its own descriptions.
Note: there is a critical direction of relationship distinction between
a virtual unit with its own descriptions and one which only references
the descriptions of real units. 'Dependant' descriptions have an outgoing
'describes' relationship to the VU whilst referenced ones have an incoming
'isDescribedBy' relationship from it. This must be observed, otherwise the
wrong description will be deleted when the VU is deleted.

cc: @lindareijnhoudt.