In-progress work on Jena SKOS import and fixing resulting fallout.
SKOS import with Jena seems to work fine (rough edges notwithstanding).
Unfortunately, Jena comes with Xerces that has subtle differences to
the reference (Sun) SaxParser that got used previously. This has broken
some of the SAX-based importers. One fix involved using the qName of
attributes instead of the localName (which is apparently blank when
not using namespaces). This leaves an outstanding problem with whitepspace
around entities and an as-yet undetermined problem with some WP2 stuff.

As a move toward a fix, the SaxXmlHandler now implements the LexicalHandler
to be notified of entity events. Hopefully this will help in determining how
whitespace is being stripped and where.