Add a sourceFileId to every DocumentsDescription, and create MaintenanceEvents for revisiondesc/creation
This uses the eadid of the eadheader, so this is a mandatory field now.
Both RevisionDesc and Creation from the eadheader are treated as MaintenanceEvents. they differ in eventType.
MaintenanceEvents should be discarded and recreated when reimporting. this has not yet been done.
MaintenanceEvents should be linked to every DocumentsDescription created from this EAD. this has not yet been done.