WIP on changing promotion to upvote/downvote.
This requires an unfortunate amount of additional methods for removing
upvotes and downvotes, as well as casting them. Promotion is determined
via a score rather than just the presence of a vote.

An odd bug as surfaced whereby adding another annotation to the fixtures
starts breaking the doc unit listing tests.