Change the identifier of an imported DocumentsDescription if it detects an existing, different one.
This detection is based on DocUnit identifier, language and sourceFileIdentifier (aka eadid).
If there is already a different description in the graph (for instance manually added, or from a different source) this description is imported next to the existing one.
The new identifier has the sourceFileIdentifier in there.
The initial use case is Yad Vashem, where there were manually added description available before the import-batch was provided.