Merge a reimported description that already had a parallel one, with the already imported one.
This might not happen that often, but the use case is like this:
DocUnit A1 has a description a1-eng. it is also included in an eadfile abc, so when imported it gets a new description with id a1-abc-eng.
when this gets re-imported it correctly merges with this a1-abc-eng instead of with a1-eng.