Add untested auto-concept-and-link generator, plus web service util refactor.
The `Linker` class automatically generates concepts from the access points
in a given repository and links them to the relevant documents with the access
point as a body.

This means that for cases where EAD has been ingested but the access points remain
"unresolved" (i.e. not clickable) we can generated a set of unambiguous concepts
automatically for a **repository-specific** vocabulary.

NB: the vocabulary has to be created first!

This commit also adds a web service interface to the linker and refactors some of the
existing utility-style web service functionality into a separate "tools" namespace.