Refactor ImporterManagers + tests: XmlImportManager -> AbstractImportManager (fixes #14)
In #14 the hierarchy of `Importer`s and `ImportManager`s is discussed and deemed strange.
For example, the `CsvImportManager` is a subclass of `XmlImportManager`. This has been
changed by renaming the `XmlImportManager` to `AbstractImportManager` - it had not much to
do with XML anyway, so this makes sense.

Some tests that referenced the `AbstractImportManager` have been refactored to use the
instance in the `AbstractImporterTest`, which now also has

```
protected Repository repository;
// Depends on fixtures
protected final String TEST_REPO = "r1";
```

and `setUp()` and `tearDown()` methods to support test setup (`@Before`) and tearing down
(`@After`).

Added or updated javadoc comments in several locations.

Remove old javadoc param, add more javadoc in Importers