Do not duplicate parent IDs in child's graph ID
Updated `joinPath(Collection<String> path)` in `IdGeneratorUtils`:
If a path part is repeated, only the last occurrence is kept.

Instead of
<http://portal.aehri.dans.knaw.nl/units/cz-002279-collection-jmp-shoah-t-collection-jmp-shoah-t-2-collection-jmp-shoah-t-2-a-collection-jmp-shoah-t-2-a-1-collection-jmp-shoah-t-2-a-1a-collection-jmp-shoah-t-2-a-1a-028>,
the URL of this item will be:
<http://portal.aehri.dans.knaw.nl/units/cz-002279-collection-jmp-shoah-t-2-a-1a-028-document-jmp-shoah-t-2-a-1a-028>.
This works even better for unit IDs without type related prefixes.

Added or updated tests to verify the behaviour in ID generation directly and by using the
shorter IDs to find nodes in the graph after import.

Added or updated javadoc comments in several locations.