Remove the parsed dates from the general unitDates property.
When creating a DocumentaryDescription, DatePeriods are created for dates that can be parsed. These dates should not be redundantly added to the DocDesc in a 'unitDates' attribute.
The way in which this is handled (in the MapImporter) could do with some improvements though. All relevant Tests have been modified. The DATE_HAS_DESCRIPTION is used to store the original unparsed datevalue.