Prevent occasional transaction failures by locking the event chain root at TX start.
Because we have a single, serialized audit event log in the graph, this
somewhat limits the extent to which concurrent graph modifications can
take place. Certain actions, such as watching items or following
users can happily work concurrently because they do not modify the event
log, but adding, modifying, or deleting data actions effectively need to hold a
lock on the event chain.

Unfortunately, Neo4j does not know that most transactions modify the
global event root, so it doesn't lock this node. This means that logged
concurrent transactions sometimes fail because they try to modify the
event root's head relationship and then discover it's already been changed
by another thread. This doesn't lead to any data corruption since the
transaction is just rolled back, but it does inconvenience the user.

This fix is an attempt to find a better balance, given that:

 - we do not ever expect a lot of write activity on the graph
 - nearly all graph modifying actions alter the global event log

What it does is override the default Blueprints/Neo4j implementation's
behaviour when starting a transaction and manually obtains a lock on the
event root node. I have tested this with a whole load of concurrent
modifying requests and it works very nicely. There is in theory a risk
of more deadlocks (since we're locking) but Neo4j should detect these
automatically.

I've not particularly happy with the precise way the root node is located,
since it depends on index implementation details that are not exposed by
lower layers, namely the name of the single manual graph index ("entities").