Fix bug with revoking all permissions on a given content type. Fixes #101.
This is a bit of a doozy. If a user has UPDATE and CREATE permissions on
a given content type it was possible to revoke one or the other, but not
both. This is because given an empty set of permissions we just ignored that
type, instead of revoking all permissions.