Improvements to frame serialization config to fix #114.
Add a way to limit the relative depth of traversal for individual
serialized relationships. Currently, fetching auto-serialized rels
recurses until a global default depth is reached. This can lead to
some pathological cases with higher depth levels, however, as well
as resulting in a lot of unnecessary data being transferred. A good
example of this is targets for links and notes: we only need the top-
level item itself in 99.9% of cases, not the whole tree and full context.
With higher depth levels, fetching all links for an item could lead to an
enormous amount of data being transferred, most of it unnecessary.

This problem is particularly acute when there's an erroneous relationship
loop, such as with #115. These loops shouldn't happen, but neither should
their occurence break things.

With this patch we can specify a relative depth limit for the target
relationship of a link (or annotation) node, so they only descend to one
level below the target (which will typically be the description for a
described item.)

Also bump version by a point since this is a fairly significant change,
albeit one that does not break any tests.

Conflicts:
	ehri-frames/src/main/java/eu/ehri/project/models/annotations/Fetch.java