Improvements to serialization of metadata values and various model cleanups.
Previously, the number of child items an `ItemHolder` has was cached in
a property. Due to the limitations of Frames, the code to do this had
to be repeated in a very ugly way. Also, while I'm not sure how much slower,
if at all, counting relationships is compared to looking up a property is,
but I doubt it's worth the additional complexity.

Also, with the caching there was inevitably issues with invalidation,
which turned out to be an actual problem.

This patch introduces a `@Meta(name)` annotation that can be affixed
to frame methods that return scalar values, which will be automatically
included in the bundle meta section.

Testing performance of full-graph serialization show neglibable performance
difference for the same functionality. Fetching watched count, however, adds
a small overhead, but one that can be elsewhere offset.