Revert merge of Sparql module with extension.
While this was not, in itself, a bad idea, it had nasty downstream
implications. Namely, the front-end test suite, which depends on the
extension module for integration tests, now had a huge extra load of
transitive dependencies (e.g. openrdf) and these happened to be ones
that SBT did not like to resolve. I think it's easier all round to
revert the module merge, but not the other fixes to the Jersey resource
path, etc.

This reverts commit 3a0df64205400769bbd0b6616f225278d68b21c1.