Rework the way events are written to the graph.
Creating an event context now does not actually change the graph. Only
when it is committed will changes take affect.

This removes the need to roll back the transaction when an idempotent
import/ingest does not end up doing any work, and potentially removes
the need to hold a lock on the event root node for the duration of an
ingest task (although this is handled separately.)

The importers now do no transaction-handling of their own, leaving that
aspect to higher levels of the stack, e.g. the command line and web service
tools.

Additional changes:

 - bumped Blueprints version to 2.6.0 (since we now have Java 7)
 - bumped version to 0.9.3