Add property accessor to base frame, plus additional cleanup.
It's easier to allow the `Frame` interface to access properties on
its underlying vertex via a forwarding method, rather than fetching
the vertex every time.

Fix import issue with an additional <p> element on the `abstract` path.
This is illegal according to the EAD schema, and does not show up in any
of the test files we have.