Add Java code to upgrade the DB format:
 - change id/type labels
 - change type values
 - ensure things like event links have id and types, in case we need
   to upgrade much old DBs
 - set labels
 - apply the schema

Whilst this code is more or less single use, it can be used to update
older DBs to the new format for inspection (or possible data recovery).

This should also, hopefully, be the last time we cannot use
Cypher for declarative DB migrations, due to the removal of legacy
indexes.