Attempt to make imports idempotent RE maintenance events.
Maintenance events were being created separately, which broke idem-
potent importing of EAD files. Now we create the given number of
MEs per-description as a dependent item, instead of a single set which
is them linked to each description.

Node counts in various tests adjusted to reflect this change (and in
some cases, fixed.)

This change can still be greatly improved due to its extreme hackiness.