Consolidate views layer into a single all powerful API class.
This is rather more consistent and easier to manage than lots of
separate classes with duplicate semantics. It will also be easier
to apply more consistency in general.

The major change to semantics is that the accessor is now an instance
variable and the item class per-method, rather than vice-versa.

Current implementations have largely been copy/pasted at the moment,
and more work needs to be done to make things nicer.