Add support for streaming GraphQL queries.
This changes the behaviour of connection pagination when the X-Stream
header is provided. Instead of limiting connections to a default of 40
and a maximum of 100 items, it uses a streaming execution strategy for
potentially unlimited connections, writing the result incrementally to
a chunked response.

This requires some not-entirely-pleasant code that duplicates quite a
lot of graphql-java's default execution code, but the utility of the
feature makes it worth it.