Add tentative support for SKOS-XL labels and better relationship sync. (#307)
This is a bit of a hack: adds RDFS subPropertyOf relations for 
skosxl:{pref,alt,hidden}Label to skos:{pref,alt,hidden}Label
and looks up the values depending on whether the object is a
literal or not.

If relationships have changed, those that have been deleted should
now be removed. This relies on inference to get symmetric narrower/
broader relationships (which is a bit sketchy.)

Also: ensure resources are closed.