Extend import logs to contain a mapping of source -> item IDs
This is part of work to improve the traceability of imported data.
Imports that come from a streamed XML body will just have "-" as the
key (standing for stdin).

As part of these changes various XML dependencies were upgraded
as well, including Jackson to 2.10.5. This necessitated a fix to
the build (using Maven shade) to exclude module-info.class files
that, since they are for JDK9+, cause an UnsupportedClassVersionError
when Neo4j loads the plugin.