Lots of refactoring to work with stricter transactional semantics
It's no longer possible to pass nodes between transactional boundaries,
which happened mostly when e.g. returning a stream of data in a separate
callback. In e.g. the web service we now have to do all interactions
with the DB within the stream handler and stuff like existence checks
separately.

A bit of other refactoring too and renaming some methods.