Fix for handling EADs with completely anonymous C-levels
This caused issues because we didn't push an ID on the stack until
we saw the <did> section, and that never happened, so the stack
got confused. This didn't manifest itself unless the import was
done in tolerant mode.

A second fix passes more info about *where* the error occurred to
the web service.

Tweaked exception handling to pass more info back to the web service
when validation errors occur.